impl: Added the ability for the train to follow the tracks
The railtracks still need to be designed and the path a bit clearer, but the ability to create a path for the train to follow is now there.
An example of a path is currently there which is followed by the train.
It needs some tweaking still, but we will be able to implement reversing and stopping trains.